Put common logic in zksync cast module and the rest on mktx and estimate submodules for zksync
